@@ -1,11 +1,12 @@
 {
   "name": "3d-game-of-life",
-  "version": "1.0.0",
+  "version": "0.1.0",
   "description": "Third dimensional game of life",
   "main": "src/index.js",
   "scripts": {
-    "build": "NODE_ENV=production webpack",
-    "server": "webpack-dev-server"
+    "build:debug": "webpack",
+    "build:relase": "NODE_ENV=production webpack",
+    "start": "webpack-dev-server -w"
   },
   "repository": {
     "type": "git",
@@ -16,19 +17,15 @@
   "bugs": {
     "url": "https://github.com/eduidl/3d-game-of-life/issues"
   },
-  "dependencies": {
-    "three": "^0.94.0",
-    "three-orbitcontrols": "^2.1.2"
-  },
   "devDependencies": {
-    "@babel/core": "^7.4.3",
-    "@babel/preset-env": "^7.4.3",
-    "babel-loader": "^8.0.4",
-    "eslint": "^5.16.0",
-    "eslint-loader": "^2.1.1",
+    "eslint": "^6.8.0",
     "html-webpack-plugin": "^3.2.0",
-    "webpack": "^4.27.1",
-    "webpack-cli": "^3.1.2",
-    "webpack-dev-server": "^3.2.1"
-  }
+    "three": "^0.112.1",
+    "ts-loader": "^6.2.1",
+    "typescript": "^3.7.4",
+    "webpack": "^4.41.5",
+    "webpack-cli": "^3.3.10",
+    "webpack-dev-server": "^3.10.1"
+  },
+  "dependencies": {}
 }
